Tabify chat settings
Creates 3 new subcomponents for chat settings, and coverts single menu dialog into a tabbed dialog
