+++ components/Chat/ChatSettings/SettingsMenu.vue
@@ -1,260 +1,68 @@
 <template>
   <div>
-    <div class="px-3">
-      <div class="d-flex">
-        <!-- Global Chat -->
-        <v-switch
-          v-model="globalChat"
-          label="Global Chat"
-          color="primary"
-          hide-details
-          dense
-          inset
-        />
-
-        <!-- Show Timestamps -->
-        <v-switch
-          v-model="showTimestamps"
-          class="ml-3"
-          label="Timestamps"
-          color="primary"
-          hide-details
-          dense
-          inset
-        />
-      </div>
-
-      <v-switch
-        v-model="receiveMentionsInLocal"
-        :disabled="globalChat"
-        class="mb-2"
-        label="Recieve @'s while in local"
-        color="primary"
-        hide-details
-        dense
-        inset
-      />
-
-      <!-- Ignore -->
-      <v-switch
-        v-model="useIgnore"
-        label="Ignore Users"
-        class="mb-2"
-        color="primary"
-        hide-details
-        dense
-        inset
-      />
-
-      <!-- Notification Sounds -->
-      <v-switch
-        v-model="notificationSound"
-        label="Notification Sounds"
-        class="mb-2"
-        color="primary"
-        hide-details
-        dense
-        inset
-      />
-
-      <!-- Autocomplete -->
-      <v-switch
-        v-model="autocomplete"
-        label="Autocomplete"
-        class="mb-2"
-        color="primary"
-        hide-details
-        dense
-        inset
-      />
-
-      <!-- Dense Chat -->
-      <v-switch
-        v-model="denseChat"
-        label="High Density"
-        class="mb-2"
-        color="primary"
-        hide-details
-        dense
-        inset
-      />
-    </div>
-
-    <v-divider/>
-
-    <v-alert
-      v-if="!ttsVoicesExist"
-      class="mt-3 mb-3"
-      type="warning"
-      text
+    <v-tabs
+      v-model="settingsTab"
+      class="elevation-2"
     >
-      No TTS voices found!
-    </v-alert>
+      <!-- Tab Slider Color -->
+      <v-tabs-slider />
 
-    <div class="pb-2" v-if="ttsVoicesExist">
+      <!-- Tabs -->
+      <v-tab
+        :href="`#chat-settings-general`"
+      >
+        General
+      </v-tab>
+      <v-tab
+        :href="`#chat-settings-tts`"
+      >
+        TTS
+      </v-tab>
+      <v-tab
+        :href="`#chat-settings-beta`"
+      >
+        Beta
+      </v-tab>
 
-      <!-- TTS Settings -->
-      <v-subheader class="overline mb-0">Text To Speech Options</v-subheader>
+      <!-- TAB: General Settings -->
+      <v-tab-item
+        value="chat-settings-general"
+        class="chat-settings-panel"
+      >
+        <chat-general-settings-panel />
+      </v-tab-item>
 
-      <div class="px-3">
-        <div class="d-flex">
-          <!-- Enable TTS -->
-          <v-switch
-            v-model="useTts"
-            :disabled="!ttsVoicesExist"
-            label="Use TTS"
-            class="mt-0 pt-0"
-            color="primary"
-            hide-details
-            dense
-            inset
-          />
+      <!-- TAB: TTS Settings -->
+      <v-tab-item
+        value="chat-settings-tts"
+        class="chat-settings-panel"
+      >
+        <chat-tts-settings-panel />
+      </v-tab-item>
 
-          <!-- Troll TTS -->
-          <v-switch
-            v-model="trollTts"
-            :disabled="!useTts"
-            label="Troll TTS"
-            class="mt-0 ml-3 pt-0"
-            color="primary"
-            hide-details
-            dense
-            inset
-          />
-	      </div>
-
-      <!-- Read username -->
-      <v-switch
-        v-model="ttsReadUsername"
-        :disabled="!useTts"
-        label="Read Username"
-        class="mb-3 pt-0"
-        color="primary"
-        hide-details
-        dense
-        inset
-      />
-      </div>
-
-      <!-- TTS Voice -->
-      <v-flex mx-3>
-        <v-select
-          v-model="ttsVoice"
-          :items="ttsVoices"
-          :disabled="!useTts"
-          label="TTS Voice"
-          style="font-size: 12px;"
-          hide-details
-        />
-      </v-flex>
-
-      <!-- TTS Rate -->
-      <v-list-item>
-        <v-slider
-          :disabled="!useTts"
-          label="Speed"
-          v-model="ttsRate"
-          class="align-center"
-          :max="20"
-          :min="1"
-          :step="0.5"
-          hide-details
-        >
-          <template #append>
-            <v-text-field
-              v-model="ttsRate"
-              class="mt-0 pt-0"
-              hide-details
-              single-line
-              type="number"
-              style="width: 50px"
-            />
-          </template>
-        </v-slider>
-      </v-list-item>
-
-      <!-- TTS Volume -->
-      <v-list-item>
-        <v-slider
-          :disabled="!useTts"
-          label="Vol."
-          v-model="ttsVolume"
-          class="align-center"
-          :max="10"
-          :min="0"
-          :step="0.5"
-          hide-details
-        />
-      </v-list-item>
-
-      <!-- TTS Timeout -->
-      <v-list-item>
-        <v-slider
-          :disabled="!useTts"
-          label="Timeout"
-          v-model="ttsTimeout"
-          class="align-center"
-          :max="60"
-          :min="0"
-          :step="0.5"
-          hide-details
-        >
-
-          <template #append>
-            <v-text-field
-              v-model="ttsTimeout"
-              class="mt-0 pt-0"
-              hide-details
-              single-line
-              type="number"
-              style="width: 50px"
-            />
-          </template>
-        </v-slider>
-      </v-list-item>
-
-    </div>
-
-    <v-divider/>
-
-    <div class="pb-5">
-      <v-subheader class="overline mb-0">
-        Experimental Options <sup style="color: red">BETA</sup>
-      </v-subheader>
-
-      <div class="px-3">
-        <v-switch
-          v-model="trackMetrics"
-          label="Keep track of chat statistics"
-          class="mb-3 mt-0"
-          color="primary"
-          hide-details
-          dense
-          inset
-        />
-
-        <v-switch
-          v-model="trackMetricsPerUser"
-          :disabled="!trackMetrics"
-          label="Track stats per-user"
-          class="mb-0"
-          color="primary"
-          hide-details
-          dense
-          inset
-        />
-      </div>
-    </div>
-
+      <!-- TAB: Beta Settings -->
+      <v-tab-item
+        value="chat-settings-beta"
+        class="chat-settings-panel"
+      >
+        <chat-beta-settings-panel />
+      </v-tab-item>
+    </v-tabs>
   </div>
 </template>
 
 <script>
-  import { mapState, mapMutations } from 'vuex'
-  import { Chat } from '@/store/chat';
+  import ChatGeneralSettingsPanel from './ChatSettingsPanel/ChatGeneralSettingsPanel';
+  import ChatTtsSettingsPanel from './ChatSettingsPanel/ChatTtsSettingsPanel';
+  import ChatBetaSettingsPanel from './ChatSettingsPanel/ChatBetaSettingsPanel';
 
   export default {
     name: 'ChatSettings',
+    components: {
+      ChatBetaSettingsPanel,
+      ChatTtsSettingsPanel,
+      ChatGeneralSettingsPanel,
+    },
 
     props: {
 
@@ -262,194 +70,27 @@
 
     data() {
       return {
-        useIgnoreList: true,
-        ttsVoices: [],
+        settingsTab: 'chat-settings-general',
       }
     },
 
     methods: {
-      ...mapMutations (Chat.namespace, {
-        setModeGlobal     : Chat.$mutations.setGlobal,
-        setModeTimestamps : Chat.$mutations.setTimestamps,
-        setUseTts         : Chat.$mutations.setUseTts,
-        setUseIgnore      : Chat.$mutations.setUseIgnore,
-        setTrollTts       : Chat.$mutations.setTrollTts,
-        setTtsRate        : Chat.$mutations.setTtsRate,
-        setTtsReadUsername: Chat.$mutations.setTtsReadUsername,
-        setTtsTimeout     : Chat.$mutations.setTtsTimeout,
-        setTtsVolume      : Chat.$mutations.setTtsVolume,
-        setTtsVoice       : Chat.$mutations.setTtsVoice,
-        setNotify         : Chat.$mutations.setNotify,
-        setAutocomplete   : Chat.$mutations.setAutocomplete,
-        setHighDensity    : Chat.$mutations.setHighDensity,
-        setReceiveMentionsInLocal : Chat.$mutations.setReceiveMentionsInLocal,
 
-        setTrackMetrics        : Chat.$mutations.setTrackMetrics,
-        setTrackMetricsPerUser : Chat.$mutations.setTrackMetricsPerUser,
-      }),
-
-      updateSettings() {
-        const settings = {};
-        this.$emit( 'input', settings );
-      },
     },
 
     computed: {
-      ...mapState (Chat.namespace, {
-        getModeGlobal     : Chat.$states.global,
-        getModeTimestamps : Chat.$states.timestamps,
-        getUseTts         : Chat.$states.useTts,
-        getUseIgnore      : Chat.$states.useIgnore,
-        getTrollTts       : Chat.$states.trollTts,
-        getTtsRate        : Chat.$states.ttsRate,
-        getTtsReadUsername: Chat.$states.ttsReadUsername,
-        getTtsTimeout     : Chat.$states.ttsTimeout,
-        getTtsVolume      : Chat.$states.ttsVolume,
-        getTtsVoice       : Chat.$states.ttsVoice,
-        getNotify         : Chat.$states.notify,
-        getAutocomplete   : Chat.$states.autocomplete,
-        getHighDensity    : Chat.$states.highDensity,
-        getReceiveMentionsInLocal : Chat.$states.receiveMentionsInLocal,
 
-        getTrackMetrics        : Chat.$states.trackMetrics,
-        getTrackMetricsPerUser : Chat.$states.trackMetricsPerUser,
-      }),
-
-      ttsVoicesExist: {
-        get () { return this.ttsVoices.length }
-      },
-
-      globalChat: {
-        set ( val ) {
-          this.setModeGlobal( val );
-          this.$analytics.logEvent( 'global_chat', { value: val } );
-        },
-        get () { return this.getModeGlobal }
-      },
-
-      showTimestamps: {
-        set ( val ) {
-          this.setModeTimestamps( val );
-          this.$analytics.logEvent( 'show_timestamps', { value: val } );
-        },
-        get () { return this.getModeTimestamps }
-      },
-
-      useTts: {
-        set ( val ) {
-          this.setUseTts( val );
-          this.$analytics.logEvent( 'use_tts', { value: val } );
-        },
-        get () { return this.getUseTts }
-      },
-
-      useIgnore: {
-        set ( val ) {
-          this.setUseIgnore( val );
-          this.$analytics.logEvent( 'use_ignore', { value: val } );
-        },
-        get () { return this.getUseIgnore }
-      },
-
-      trollTts: {
-        set ( val ) {
-          this.setTrollTts( val );
-          this.$analytics.logEvent( 'allow_troll_tts', { value: val } );
-        },
-        get () { return this.getTrollTts }
-      },
-
-      ttsRate: {
-        set ( val ) {
-          this.setTtsRate( val );
-          this.$analytics.logEvent( 'tts_rate', { value: val } );
-        },
-        get () { return this.getTtsRate }
-      },
-
-      ttsReadUsername: {
-        set ( val ) {
-          this.setTtsReadUsername( val );
-          this.$analytics.logEvent( 'tts_read_username', { value: val } );
-        },
-        get () { return this.getTtsReadUsername }
-      },
-
-      ttsTimeout: {
-        set ( val ) {
-          this.setTtsTimeout( val );
-          this.$analytics.logEvent( 'tts_timeout', { value: val } );
-        },
-        get () { return this.getTtsTimeout }
-      },
-
-      ttsVolume: {
-        set ( val ) {
-          this.setTtsVolume( val );
-          this.$analytics.logEvent( 'tts_volume', { value: val } );
-        },
-        get () { return this.getTtsVolume }
-      },
-
-      ttsVoice: {
-        set ( val ) {
-          this.setTtsVoice( val );
-        },
-        get () { return this.getTtsVoice }
-      },
-
-      notificationSound: {
-        set ( val ) {
-          this.setNotify( val );
-          this.$analytics.logEvent( 'chat_notifications', { value: val } );
-        },
-        get () { return this.getNotify }
-      },
-
-      autocomplete: {
-        set ( val ) {
-          this.setAutocomplete( val );
-          this.$analytics.logEvent( 'chat_autocomplete', { value: val } );
-        },
-        get () { return this.getAutocomplete }
-      },
-
-      denseChat: {
-        set ( val ) {
-          this.setHighDensity( val );
-          this.$analytics.logEvent( 'chat_high_density', { value: val } );
-        },
-        get () { return this.getHighDensity }
-      },
-
-      receiveMentionsInLocal: {
-        set ( val ) {
-          this.setReceiveMentionsInLocal( val );
-          this.$analytics.logEvent( 'chat_recieve_mentions_in_local', { value: val } );
-        },
-        get () { return this.getReceiveMentionsInLocal }
-      },
-
-      trackMetrics: {
-        set ( val ) {
-          this.setTrackMetrics( val );
-          this.$analytics.logEvent( 'chat_track_metrics', { value: val } );
-        },
-        get () { return this.getTrackMetrics }
-      },
-
-      trackMetricsPerUser: {
-        set ( val ) {
-          this.setTrackMetricsPerUser( val );
-          this.$analytics.logEvent( 'chat_track_metrics_per_user', { value: val } );
-        },
-        get () { return this.getTrackMetricsPerUser }
-      },
     },
 
     async mounted () {
-      this.ttsVoices = speechSynthesis.getVoices().map( ( voice, index ) => ({ text: voice.name, value: index }));
-      speechSynthesis.onvoiceschanged = () => this.ttsVoices = speechSynthesis.getVoices().map( ( voice, index ) => ({ text: voice.name, value: index }) );
+
     },
   }
 </script>
+
+<style lang="scss">
+  .chat-settings-panel {
+    height: 20rem;
+    overflow: auto;
+  }
+</style>
