+++ components/Chat/ChatSettings/index.vue
@@ -3,11 +3,14 @@
     v-model="showChatSettings"
     :close-on-content-click="false"
     transition="slide-x-transition"
-    :max-width="320"
+    :max-width="350"
     max-height="90vh"
     top
     right
     offset-y
+    eager
+    origin="bottom left"
+    ref="chatSettingsMenu"
   >
     <template #activator="{ on: menu }">
       <v-tooltip top :open-delay="500">
@@ -20,26 +23,36 @@
       </v-tooltip>
     </template>
 
-    <v-card class="mb-0" color="grey darken-4">
-      <!-- Title Bar -->
-      <v-sheet
-        tile
-        color="primary"
-        class="d-flex align-center justify-space-between pl-2"
+    <div class="mb-3">
+      <v-card
+        color="grey darken-4"
       >
-        <h5 class="black--text body-2">Chat Settings</h5>
-        <v-btn
-          color="black"
-          text
-          icon
-          pa-0
-          @click="close"
+        <!-- Title Bar -->
+        <v-sheet
+          tile
+          color="primary"
+          class="d-flex align-center justify-space-between pl-2"
         >
-          <v-icon color="black">close</v-icon>
-        </v-btn>
-      </v-sheet>
-      <settings-menu />
-    </v-card>
+          <h5 class="black--text body-2 font-weight-medium">
+            Chat Settings
+          </h5>
+          <v-btn
+            color="black"
+            text
+            icon
+            pa-0
+            @click="close"
+          >
+            <v-icon color="black">close</v-icon>
+          </v-btn>
+        </v-sheet>
+
+        <!-- Tabbed Settings -->
+        <settings-menu @change="onTabChange" />
+
+      </v-card>
+    </div>
+
   </v-menu>
 </template>
 
@@ -63,6 +76,10 @@
       close () {
         this.showChatSettings = false;
       },
+
+      onTabChange () {
+        this.$nextTick( () => this.$refs.chatSettingsMenu.onResize() );
+      }
     },
   };
 </script>
